requirements.txt
torch~=1.10.2
bitstring~=4.2.3
pytorch-lightning~=1.9.3
torchmetrics~=0.11.4
torchvision~=0.11.1
pandas~=2.0.3
numpy~=1.24.4
pillow~=10.4.0
tqdm~=4.66.5
pyldpc~=0.7.9
hydra-core~=1.3.2
omegaconf~=2.3.0
matplotlib~=3.7.5
